requirements.txt
anyio==4.8.0
APScheduler==3.6.3
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
boto3==1.39.6
botocore==1.39.6
cachetools==4.2.2
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.44.1
decorator==5.1.1
dj-database-url==3.0.1
Django==5.1.4
django-cloudinary-storage==0.3.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-js-asset==3.1.2
django-phonenumber-field==8.1.0
django-storages==1.14.6
django-tinymce==4.1.0
executing==2.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
ipython==8.30.0
jedi==0.19.2
jmespath==1.0.1
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
phonenumbers==9.0.7
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.3
pytz==2024.2
requests==2.32.3
rfc3986==1.5.0
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.1
traitlets==5.14.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
wcwidth==0.2.13
whitenoise==6.9.0
